refactor: share resolveVariable scope walk
